refactor: move dumbo, mmds and logger to vmm (4/11)
Update dumbo to be a module.

Note: This is part of commits having the same title.
