+++ src/vmm/src/dumbo/mod.rs
@@ -12,12 +12,12 @@ use std::ops::Index;
 
 use utils::net::mac::MacAddr;
 
-pub use crate::pdu::arp::{EthIPv4ArpFrame, ETH_IPV4_FRAME_LEN};
-pub use crate::pdu::ethernet::{
+pub use crate::dumbo::pdu::arp::{EthIPv4ArpFrame, ETH_IPV4_FRAME_LEN};
+pub use crate::dumbo::pdu::ethernet::{
     EthernetFrame, ETHERTYPE_ARP, ETHERTYPE_IPV4, PAYLOAD_OFFSET as ETHERNET_PAYLOAD_OFFSET,
 };
-pub use crate::pdu::ipv4::{IPv4Packet, PROTOCOL_TCP, PROTOCOL_UDP};
-pub use crate::pdu::udp::{UdpDatagram, UDP_HEADER_SIZE};
+pub use crate::dumbo::pdu::ipv4::{IPv4Packet, PROTOCOL_TCP, PROTOCOL_UDP};
+pub use crate::dumbo::pdu::udp::{UdpDatagram, UDP_HEADER_SIZE};
 
 /// Represents a generalization of a borrowed `[u8]` slice.
 #[allow(clippy::len_without_is_empty)]
