+++ src/vmm/src/dumbo/pdu/ethernet.rs
@@ -9,7 +9,7 @@ use std::result::Result;
 
 use super::bytes::{InnerBytes, NetworkBytes, NetworkBytesMut};
 use super::Incomplete;
-use crate::MacAddr;
+use crate::dumbo::MacAddr;
 
 const DST_MAC_OFFSET: usize = 0;
 const SRC_MAC_OFFSET: usize = 6;
